Fixed issue #2
It affects 64-bit Linux including aarch64, mips64, and x86-64 Android build.

+++ src/md5.h
@@ -98,7 +98,7 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 // UINT4 defines a four byte word
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;	// Original: typedef unsigned long int UINT4;
 
 
 // convenient object that wraps
